feedback: reread connstrings
